pyproject.toml
[project]
name = "oss20b"
version = "0.1.0"
description = "FastAPI LLM based on Ollama"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.99.4",
    "pillow>=11.3.0",
    "pypdf2>=3.0.1",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.6",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.1",
]
